requirements.txt
streamlit
plotly
pandas
reportlab
python-docx
lxml
Pillow
matplotlib
plotly[kaleido]
